NF: Move CustomStudyDialog to its own package
