+++ AnkiDroid/src/main/java/com/ichi2/anki/StudyOptionsActivity.java
@@ -21,9 +21,8 @@ import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
 
-import com.ichi2.anim.ActivityTransitionAnimation;
 import com.ichi2.anki.StudyOptionsFragment.StudyOptionsListener;
-import com.ichi2.anki.dialogs.CustomStudyDialog;
+import com.ichi2.anki.dialogs.customstudy.CustomStudyDialog;
 import com.ichi2.widget.WidgetStatus;
 
 import timber.log.Timber;
