+++ AnkiDroid/src/test/java/com/ichi2/anki/dialogs/CustomStudyDialogTest.java
@@ -24,9 +24,9 @@ import com.ichi2.anki.AnkiActivity;
 import com.ichi2.anki.CardTemplateBrowserAppearanceEditor;
 import com.ichi2.anki.R;
 import com.ichi2.anki.RobolectricTest;
-import com.ichi2.anki.dialogs.CustomStudyDialog.CustomStudyListener;
+import com.ichi2.anki.dialogs.customstudy.CustomStudyDialog;
+import com.ichi2.anki.dialogs.customstudy.CustomStudyDialog.CustomStudyListener;
 import com.ichi2.libanki.Deck;
-import com.ichi2.utils.JSONObject;
 
 import org.hamcrest.Matchers;
 import org.junit.Ignore;
@@ -39,7 +39,6 @@ import androidx.fragment.app.testing.FragmentScenario;
 import androidx.lifecycle.Lifecycle;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import static com.ichi2.libanki.Consts.DECK_DYN;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.core.IsNull.notNullValue;
